Don't cache responses that don't have a stale-while-revalidate cache-control header set
The whole point of this proxy is to implement swr. It's purpose is not a (ideally blazing fast) CDN proxy.

Our file-system based cache backend is quite slow, bypassing is usually (much) faster.

If you need a fast CDN cache, use a CDN. If the CDN doesn't support swr, use this project additionally.